Remove legacy code.
