+++ osu.Game.Rulesets.Karaoke.Tests/Edit/TestSceneLayout.cs
@@ -1,0 +1,53 @@
+﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using NUnit.Framework;
+using osu.Framework.Allocation;
+using osu.Framework.Graphics;
+using osu.Game.Rulesets.Edit;
+using osu.Game.Rulesets.Karaoke.Beatmaps;
+using osu.Game.Rulesets.Karaoke.Edit.Layout;
+using osu.Game.Rulesets.Karaoke.Skinning;
+using osu.Game.Rulesets.Karaoke.Tests.Beatmaps;
+using osu.Game.Screens.Edit;
+using osu.Game.Skinning;
+using osu.Game.Tests.Visual;
+
+namespace osu.Game.Rulesets.Karaoke.Tests.Edit
+{
+    [TestFixture]
+    public class TestSceneLayout : EditorClockTestScene
+    {
+        private readonly KaraokeLayoutTestSkin skin = new KaraokeLayoutTestSkin();
+
+        [Cached(typeof(EditorBeatmap))]
+        [Cached(typeof(IBeatSnapProvider))]
+        private readonly EditorBeatmap editorBeatmap;
+
+        public TestSceneLayout()
+        {
+            var beatmap = new TestKaraokeBeatmap(null);
+            var karaokeBeatmap = new KaraokeBeatmapConverter(beatmap, new KaraokeRuleset()).Convert() as KaraokeBeatmap;
+            editorBeatmap = new EditorBeatmap(karaokeBeatmap);
+        }
+
+        [BackgroundDependencyLoader]
+        private void load()
+        {
+            Beatmap.Value = CreateWorkingBeatmap(editorBeatmap.PlayableBeatmap);
+            Child = new SkinProvidingContainer(skin)
+            {
+                RelativeSizeAxes = Axes.Both,
+                Child = new LayoutScreen(),
+            };
+        }
+
+        public class KaraokeLayoutTestSkin : KaraokeLegacySkinTransformer
+        {
+            public KaraokeLayoutTestSkin()
+                : base(null)
+            {
+            }
+        }
+    }
+}
